Aplica format en algunas de las clases nuevas

+++ sic-api/src/main/java/sic/modelo/BusquedaFacturaCompraCriteria.java
@@ -18,7 +18,7 @@ public class BusquedaFacturaCompraCriteria {
     private boolean buscaPorProveedor;
     private Proveedor proveedor;
     private boolean buscaPorNumeroFactura;
-    private int numSerie;
+    private long numSerie;
     private long numFactura;
     private boolean buscarSoloInpagas;
     private boolean buscaSoloPagadas;
