+++ sic-api/src/main/java/sic/modelo/BusquedaFacturaVentaCriteria.java
@@ -22,7 +22,7 @@ public class BusquedaFacturaVentaCriteria {
     private boolean buscaUsuario;
     private Usuario usuario;
     private boolean buscaPorNumeroFactura;
-    private int numSerie;
+    private long numSerie;
     private long numFactura;
     private boolean buscarPorPedido;
     private long nroPedido;
